requirements.txt
faiss-cpu==1.7.4

sentence-transformers==2.6.1
transformers==4.37.2
torch==2.1.2
huggingface_hub==0.20.3

numpy==1.24.4
scipy==1.10.1
pandas==2.0.3

python-docx==1.1.0
pypdf==3.17.4
openpyxl==3.1.2

langdetect==1.0.9
nltk==3.8.1

orjson==3.9.15
tqdm==4.66.1
email-validator==2.1.0.post1
chardet==5.2.0

pydantic==1.10.13

pytest==7.4.4
